+++ zenoh/src/api/builders/mod.rs
@@ -13,6 +13,7 @@
 //
 
 pub(crate) mod info;
+pub(crate) mod matching_listener;
 pub(crate) mod publisher;
 pub(crate) mod query;
 pub(crate) mod queryable;
